Moved morpc.frictionless.md5() to morpc.md5(). Application is not specific to frictionless.
